app/requirements.txt
torch
tqdm
torchvision
pillow
albumentations
torchaudio
opencv-python
streamlit